Move into shared folder and add welsh translations
