+++ VideoWeb/VideoWeb/ClientApp/src/app/shared/confirmation/yes-no-popup-base.component.scss
