+++ VideoWeb/VideoWeb/ClientApp/src/app/waiting-space/confirmation/confirm-close-hearing-popup.component.ts
@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
-import { YesNoPopupBaseDirective } from './yes-no-popup-base.component';
+import { YesNoPopupBaseDirective } from '../../shared/confirmation/yes-no-popup-base.component';
 import { FocusService } from 'src/app/services/focus.service';
 
 @Component({
     selector: 'app-confirm-close-hearing-popup',
     templateUrl: './confirm-close-hearing-popup.component.html',
-    styleUrls: ['./yes-no-popup-base.component.scss']
+    styleUrls: ['../../shared/confirmation/yes-no-popup-base.component.scss']
 })
 export class ConfirmCloseHearingPopupComponent extends YesNoPopupBaseDirective {
     constructor(protected focusService: FocusService) {
