+++ VideoWeb/VideoWeb/ClientApp/src/app/waiting-space/confirmation/confirm-join-hearing-popup.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, ViewChild } from '@angular/core';
-import { YesNoPopupBaseDirective } from './yes-no-popup-base.component';
+import { YesNoPopupBaseDirective } from '../../shared/confirmation/yes-no-popup-base.component';
 import { FEATURE_FLAGS, LaunchDarklyService } from 'src/app/services/launch-darkly.service';
 import { MuteMicrophoneComponent } from '../mute-microphone/mute-microphone.component';
 import { FocusService } from 'src/app/services/focus.service';
@@ -7,7 +7,7 @@ import { FocusService } from 'src/app/services/focus.service';
 @Component({
     selector: 'app-confirm-join-hearing-popup',
     templateUrl: './confirm-join-hearing-popup.component.html',
-    styleUrls: ['./yes-no-popup-base.component.scss']
+    styleUrls: ['../../shared/confirmation/yes-no-popup-base.component.scss']
 })
 export class ConfirmJoinHearingPopupComponent extends YesNoPopupBaseDirective {
     @ViewChild(MuteMicrophoneComponent) muteMicrophoneForm: MuteMicrophoneComponent;
