+++ VideoWeb/VideoWeb/ClientApp/src/app/waiting-space/confirmation/warn-join-hearing-popup.component.ts
@@ -1,10 +1,10 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
-import { YesNoPopupBaseDirective } from './yes-no-popup-base.component';
+import { YesNoPopupBaseDirective } from '../../shared/confirmation/yes-no-popup-base.component';
 
 @Component({
     selector: 'app-warn-join-hearing-popup',
     templateUrl: './warn-join-hearing-popup.component.html',
-    styleUrls: ['./yes-no-popup-base.component.scss']
+    styleUrls: ['../../shared/confirmation/yes-no-popup-base.component.scss']
 })
 export class WarnJoinHearingPopupComponent extends YesNoPopupBaseDirective {
     @Input() hearingStartTime: Date;
